--- .storybook/addons.js
@@ -1,6 +1,0 @@
-/* eslint-disable import/no-extraneous-dependencies */
-// The order of those imports will be the same as the order of tabs in addons pane
-import '@storybook/addon-actions/register';
-import '@storybook/addon-knobs/register';
-import '@storybook/addon-a11y/register';
-import '@storybook/addon-notes/register';
