+++ package.json
@@ -36,12 +36,14 @@
   "devDependencies": {
     "@babel/core": "^7.7.7",
     "@babel/plugin-syntax-dynamic-import": "^7.7.4",
-    "@storybook/addon-a11y": "^5.2.8",
-    "@storybook/addon-actions": "^5.2.8",
-    "@storybook/addon-knobs": "^5.2.8",
-    "@storybook/addon-notes": "^5.2.8",
-    "@storybook/html": "^5.2.8",
-    "@storybook/theming": "^5.2.8",
+    "@storybook/addon-a11y": "^5.3.0-rc.0",
+    "@storybook/addon-actions": "^5.3.0-rc.0",
+    "@storybook/addon-knobs": "^5.3.0-rc.0",
+    "@storybook/addon-notes": "^5.3.0-rc.0",
+    "@storybook/addons": "^5.3.0-rc.0",
+    "@storybook/csf": "^0.0.1",
+    "@storybook/html": "^5.3.0-rc.0",
+    "@storybook/theming": "^5.3.0-rc.0",
     "babel-loader": "^8.0.6",
     "babel-plugin-inline-svg": "^1.0.1",
     "babel-plugin-template-html-minifier": "^3.1.0",
